requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    ./compile-requirements.sh
#
-e git+https://github.com/jaywink/federation.git@2a02f866aad037495b9d56f1a6c3e44a5e04c20a#egg=federation==0.15.0.1
certifi==2018.1.18        # via requests
chardet==3.0.4            # via requests
circus==0.14.0
click==6.7                # via rq
croniter==0.3.20          # via rq-scheduler
cssselect==1.0.3
decorator==4.2.1          # via ipython, traitlets
dirty-validators==0.4.0
django-countries==5.2
django-enumfields==0.10.0
django-environ==0.4.4
django-extensions==2.0.6
django-redis==4.9.0
django-rq==1.0.1
django==2.0.3
idna==2.6                 # via requests
iowait==0.2               # via circus
ipython-genutils==0.2.0   # via traitlets
ipython==6.2.1
isodate==0.6.0            # via python-xrd
jedi==0.11.1              # via ipython
jsonschema==2.6.0
lxml==4.2.0
parso==0.1.1              # via jedi
pexpect==4.4.0            # via ipython
pickleshare==0.7.4        # via ipython
prompt-toolkit==1.0.15    # via ipython
psutil==5.4.3             # via circus
psycopg2-binary==2.7.4
ptyprocess==0.5.2         # via pexpect
pycrypto==2.6.1
pygments==2.2.0           # via ipython
python-dateutil==2.7.0    # via croniter
python-xrd==0.1
pytz==2018.3              # via django
pyzmq==17.0.0             # via circus
raven==6.6.0
redis==2.10.6
requests==2.18.4
rq-scheduler==0.8.2
rq==0.10.0
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via django-environ, django-extensions, isodate, prompt-toolkit, python-dateutil, traitlets
tornado==4.5.3
traitlets==4.3.2          # via ipython
urllib3==1.22             # via requests
uwsgi==2.0.17
wcwidth==0.1.7            # via prompt-toolkit
whitenoise==3.3.1
